Updated functions to be pure
In response to https://github.com/myxie/shadow/issues/20, functions
should all be 'pure' - that is, have a return value, and not
mutate inputs as a side effect.

This has affected the ranking functions used in heft() and pheft() most
significantly, as they have been changed from a recursive to
iterative implementation.