Changes to heuristic.py and Solution/Allocation classes
Moving towards deprecating Task objects storing machine allocations,
start and finish times;
this leads to issues when updating solutions whilst not updating the
Workflow Task object as well; it also means that the workflow object is
changed without notifying the user, so this has been changed.

It also makes sense for the Allocation object to store details of when
a task starts and stops, as these values are allocation-dependent.

This has allowed the heuristc.py file to be more 'pure' with its use
of functions, and all functions are (hopefully) no longer changing the
state of parameter variables.

This _will_ break metaheuristic functionality for the time being, but
moving forward I expect it to significantly improve the usability
of these classes.