Improving the diversity of the solutions generated by changing the way we generate random functions.
A big TODO is now to keep track of the generated values for
a particular seed, as the existing test-cases are 'broken'
due to changing the randomly generated values.

We may be able to update some of the test cases to make sure
that the function works for a 'given' solution - i.e. we could
pickle one of the solution objects and use that to ensure backwards compatibility when refactoring.