Speed improvements for calculation costs.
Previously we were pre-calculating all task costs on
machines and then accessing through calculated_runtime.

When running with large files, this was ridiculously wasteful and would cause SHADOW to hang.

Now it can process an example skaworkflow (>4400 tasks) on a HPC environment (~900 machines) in ~1minute.

FCFS was also incorrect, so we fixed that up.